% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AICc.R
\name{AICc}
\alias{AICc}
\title{Akaike's Information Criterion (corrected)}
\usage{
AICc(mod)
}
\arguments{
\item{mod}{A model object}
}
\value{
A computed value for AICc
}
\description{
This function computes the AICc for a given model for which a basic AIC can
be obtained
}
\details{
The AICc is a small-sample corrected AIC computed using the following
formula: \deqn{AIC_c = AIC + \frac{2k^2 + 2k}{n-k-1}}
where k is the number of parameters in the model and n is the sample size. This
has the effect of choosing a more parsimonious model to avoid overfitting at
small sample sizes, but \eqn{AIC_c \rightarrow AIC} as \eqn{n \rightarrow \infty}
}
