% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B_sim.R
\name{B_sim}
\alias{B_sim}
\title{Simulate average unfished biomass}
\usage{
B_sim(
  recruits,
  spawners,
  strata = NA,
  mean_wt,
  maturity,
  S,
  num_years = 2500,
  model_type = c("empirical", "empirical2"),
  lnorm_err = NULL
)
}
\arguments{
\item{recruits}{numeric vector of estimates of annual recruits (millions)}

\item{spawners}{numeric vector of estimates for spawning biomass three years
prior (tonnes)}

\item{strata}{factor vector giving the strata for the empirical spawner-recruit
relationship}

\item{mean_wt}{numeric vector of mean weight-at-age for ages 3 to 8+}

\item{maturity}{numeric vector of estimated maturity-at-age for ages 3 to 8+}

\item{S}{estimated constant survival rate for all age classes}

\item{num_years}{Integer; the number of years to simulate to}

\item{model_type}{The type of spawner-recruit model. If empirical is chosen,
then age-3 recruits are randomly sampled from strata based on spawning biomass
three years prior. If some other character string is given, then the string
must correspond to the name of a functional spawner-recruit model made elsewhere}

\item{lnorm_err}{A lognormal error term to be used, if desired, if using a
functional spawner-recruit model.}
}
\value{
A data frame made up of the following vectors
\itemize{
\item \strong{YEAR:} Index of simulated year
\item \strong{B:} Simulated biomass (tons) for that year
\item \strong{B_cumavg:} The cumulative average biomass (tons) up to that year
\item \strong{AUB.25:} The estimated 25 percent management threshold for entire series
\item \strong{AUB:} The estimated average unfished biomass
}
}
\description{
This function simulates the average unfished biomass of herring using a
user-specified spawner-recruit relationship and data including weight-at-age,
proportion of stock that are mature, and a constant survival rate.
}
