% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B_sim_functional.R
\name{B_sim_functional}
\alias{B_sim_functional}
\title{Simulate average unfished biomass using functional}
\usage{
B_sim_functional(
  N_t = NULL,
  recruits,
  spawners,
  mean_wt,
  maturity,
  S,
  model_type,
  num_years = 5000,
  max_iters = 10000
)
}
\arguments{
\item{N_t}{initial data frame of herring numbers-at-age.
Generated by \code{\link[=B_sim_init]{B_sim_init()}}}

\item{recruits}{numeric vector of estimates of annual recruits (millions)}

\item{spawners}{numeric vector of estimates for spawning biomass three years
prior (tonnes)}

\item{mean_wt}{numeric vector of mean weight-at-age for ages 3 to 8+}

\item{maturity}{numeric vector of estimated maturity-at-age for ages 3 to 8+}

\item{S}{estimated constant survival rate for all age classes}

\item{model_type}{Character string corresponding to the name of a
stock-recruit model existing in the users global environment.}

\item{num_years}{Integer; the number of years to simulate to}

\item{max_iters}{Integer; the maximum number of years to simulate before throwing
an error if convergence in cumulative mean series is not achieved. Must be
greater than \code{num_years}}
}
\value{
A data frame made up of the following vectors
\itemize{
\item \strong{YEAR:} Index of simulated year
\item \strong{B:} Simulated biomass (tons) for that year
\item \strong{B_cumavg:} The cumulative average biomass (tons) up to that year
\item \strong{AUB.25:} The estimated 25 percent management threshold for entire series
\item \strong{AUB:} The estimated average unfished biomass
}
}
\description{
This function simulates the average unfished biomass of herring using a
user-specified functional spawner-recruit relationship such as ricker or
beverton-holt. Requires data including weight-at-age, proportion of stock
that are mature, and a constant survival rate.
}
