% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B_sim_init.R
\name{B_sim_init}
\alias{B_sim_init}
\title{Initialize data frame to simulate unfished biomass}
\usage{
B_sim_init(recruits, S)
}
\arguments{
\item{recruits}{numeric vector of estimates of annual recruits (millions)}

\item{S}{estimated constant survival rate for all age classes}
}
\value{
A data frame made up of the following vectors
\itemize{
\item \strong{age_3:} vector of age 3 recruits
\item \strong{age_4:} vector of age 4 herring
\item \strong{age_5:} vector of age 5 herring
\item \strong{age_6:} vector of age 6 herring
\item \strong{age_7:} vector of age 7 herring
\item \strong{age_8:} plus group vector of age 8+ herring
}
}
\description{
This function initializes a data frame to be used by any of \code{\link[=B_sim_empirical]{B_sim_empirical()}},
\code{\link[=B_sim_lnRecruitment]{B_sim_lnRecruitment()}}, or \code{\link[=B_sim_functional]{B_sim_functional()}}. The data frame is mostly \code{NA}'s
to be filled by other functions with simulated herring counts; the first 8 years
of age-3 recruit counts are sampled from the provided vector of \code{recruits}. Then
counts for every year thereafter as the \code{recruits} age are computed from the
constant survival rate \code{S}.
}
