% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B_sim_lnRecruitment.R
\name{B_sim_lnRecruitment}
\alias{B_sim_lnRecruitment}
\title{Simulate average unfished biomass using lognormal recruitment}
\usage{
B_sim_lnRecruitment(
  N_t,
  recruits,
  spawners,
  strata = NA,
  mean_wt,
  maturity,
  S,
  num_years = 2500
)
}
\arguments{
\item{N_t}{initial data frame of herring numbers-at-age.
Generated by \code{\link[=B_sim_init]{B_sim_init()}}}

\item{recruits}{numeric vector of estimates of annual recruits (millions)}

\item{spawners}{numeric vector of estimates for spawning biomass three years
prior (metric tons)}

\item{strata}{factor vector giving the strata for the empirical spawner-recruit
relationship}

\item{mean_wt}{numeric vector of mean weight-at-age for ages 3 to 8+}

\item{maturity}{numeric vector of estimated maturity-at-age for ages 3 to 8+}

\item{S}{estimated constant survival rate for all age classes}

\item{num_years}{Integer; the number of years to simulate to}
}
\value{
A data frame made up of the following vectors
\itemize{
\item \strong{YEAR:} Index of simulated year
\item \strong{B:} Simulated biomass (tons) for that year
\item \strong{B_cumavg:} The cumulative average biomass (tons) up to that year
\item \strong{AUB.25:} The estimated 25 percent management threshold for entire series
\item \strong{AUB:} The estimated average unfished biomass
}
}
\description{
This function simulates the average unfished biomass of herring using a
lognormal recruitment strategy and data including
weight-at-age, proportion of stock that are mature, and a constant survival
rate.
}
\details{
Recruitment is treated as a quasi-spawner-independent process. The spawner-
recruit estimates may or may not be stratified. Within each strata, the mean
recruitment multiplied by a log-normal error, where the variance term is
computed from the data within each strata.
}
\references{
\insertRef{methot2011}{herringSim}
}
