% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caa_asa2021.R
\docType{data}
\name{caa_asa2021}
\alias{caa_asa2021}
\title{Catch-at-age 1979-2021}
\format{
An object of class \code{data.frame} with 46 rows and 7 columns.
}
\usage{
caa_asa2021
}
\description{
Catch-at-age for Sitka Sound herring fishery from ASA 2022-forecast
}
\keyword{datasets}
