% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caa_asa2023_forecast.R
\docType{data}
\name{caa_asa2023_forecast}
\alias{caa_asa2023_forecast}
\title{Catch-at-age 1976-2022}
\format{
An object of class \code{data.frame} with 47 rows and 7 columns.
}
\usage{
caa_asa2023_forecast
}
\description{
Catch-at-age for Sitka Sound herring fishery from ASA 2023-forecast
}
\keyword{datasets}
