% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cols.R
\docType{data}
\name{cols}
\alias{cols}
\title{herringSim Color Palette}
\format{
An object of class \code{character} of length 8.
}
\usage{
cols
}
\description{
Custom color palatte used in herringSim graphics and analysis
}
\keyword{datasets}
