% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_cor.R
\name{exp_cor}
\alias{exp_cor}
\title{Exploratory Correlation Measure}
\usage{
exp_cor(x, y, method = "kendall", trim_pct = 0.9)
}
\arguments{
\item{x}{numeric vector for which correlation measure is desired.}

\item{y}{numeric vector for which correlation measure is desired. This vector
will be trimmed of outliers, hence it should correspond to a response variable
of interest.}

\item{method}{correlation method corresponding to those accepted by \code{cor} and
\code{cor.test}, i.e. one of \code{pearson}, \code{kendall}, or \code{spearman}.}

\item{trim_pct}{percentage of data to keep when trimming.}
}
\value{
Exploratory correlation measure. Values outside of \eqn{\pm} 3 correspond to
p-values significant at an \eqn{\alpha}-level of 0.05
}
\description{
This function computes the exploratory correlation measure as described in
Williams & Quinn (2000).
}
\references{
\insertRef{williamsquinn2000}{herringSim}
}
