% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_convergence.R
\name{find_convergence}
\alias{find_convergence}
\title{Find Series convergence}
\usage{
find_convergence(x, n = 100, limit = 0.05)
}
\arguments{
\item{x}{A numeric vector}

\item{n}{A positive integer; defines both window size and how many window RMSD's
must fall below the limit for the convergence criteria to be considered satisfied}

\item{limit}{A positive numeric; the threshold value under which RMSD must fall
in order to satisfy the convergence criteria}
}
\value{
The index of \code{x} after which the convergence criteria is met
}
\description{
This function decides when a series converges using two user-specified
criteria. Convergence is declared when the window root mean squared deviation
fall below some limiting value for n consecutive elements in the series.
}
