% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_intercept.R
\name{fit_intercept}
\alias{fit_intercept}
\title{Fit intercept-only stock-recruit model}
\usage{
fit_intercept(
  recruits,
  spawners,
  start,
  errorStructure = c("additive", "multiplicative")
)
}
\arguments{
\item{recruits}{numeric vector of estimates of annual recruits (millions)}

\item{spawners}{numeric vector of estimates for spawning biomass three years
prior (tonnes)}

\item{start}{numeric vector of starting values for model parameters}

\item{errorStructure}{string specifying the error structure of model}
}
\value{
A named list
\itemize{
\item \strong{coef:} Estimated model parameters
\item \strong{mod:} Function implementing model with estimated parameters
\item \strong{logLik:} Log-likelihood for model
}
}
\description{
This function uses maximum likelihood to fit an intercept-only stock-recruitment
model.
}
