% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maturity_asa1996.R
\docType{data}
\name{maturity_asa1996}
\alias{maturity_asa1996}
\title{Maturity-at-age 1973-1996}
\format{
An object of class \code{data.frame} with 23 rows and 6 columns.
}
\usage{
maturity_asa1996
}
\description{
Proportional maturity-at-age estimates from 1973 to 1996 from ASA 1996.
}
\keyword{datasets}
