% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maturity_asa2021.R
\docType{data}
\name{maturity_asa2021}
\alias{maturity_asa2021}
\title{Maturity-at-age 1979-2021}
\format{
An object of class \code{data.frame} with 46 rows and 6 columns.
}
\usage{
maturity_asa2021
}
\description{
Proportional maturity-at-age estimates from 1979 to 2021 from ASA 2022-forecast.
}
\keyword{datasets}
