% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maturity_asa2023_forecast.R
\docType{data}
\name{maturity_asa2023_forecast}
\alias{maturity_asa2023_forecast}
\title{Maturity-at-age 1976-2022}
\format{
An object of class \code{data.frame} with 47 rows and 6 columns.
}
\usage{
maturity_asa2023_forecast
}
\description{
Proportional maturity-at-age estimates from 1976 to 2022 from ASA 2023-forecast.
}
\keyword{datasets}
