% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_wt_asa1996.R
\docType{data}
\name{mean_wt_asa1996}
\alias{mean_wt_asa1996}
\title{Weight-at-age 1973-1996}
\format{
An object of class \code{data.frame} with 23 rows and 6 columns.
}
\usage{
mean_wt_asa1996
}
\description{
Weight-at-age (grams) estimates from 1973 to 1996 from age-weight-length sampling
in Sitka sound.
}
\keyword{datasets}
