% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_wt_asa2021.R
\docType{data}
\name{mean_wt_asa2021}
\alias{mean_wt_asa2021}
\title{Weight-at-age 1979-2021}
\format{
An object of class \code{data.frame} with 46 rows and 6 columns.
}
\usage{
mean_wt_asa2021
}
\description{
Weight-at-age (grams) estimates from 1979 to 2021 from age-weight-length sampling
in Sitka sound.
}
\keyword{datasets}
