% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_wt_asa2023_forecast.R
\docType{data}
\name{mean_wt_asa2023_forecast}
\alias{mean_wt_asa2023_forecast}
\title{Weight-at-age 1976-2022}
\format{
An object of class \code{data.frame} with 47 rows and 6 columns.
}
\usage{
mean_wt_asa2023_forecast
}
\description{
Weight-at-age (grams) estimates from 1976 to 2022 from age-weight-length sampling
in Sitka sound.
}
\keyword{datasets}
