% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naa_asa2021.R
\docType{data}
\name{naa_asa2021}
\alias{naa_asa2021}
\title{Numbers-at-age 1979-2021}
\format{
An object of class \code{data.frame} with 46 rows and 7 columns.
}
\usage{
naa_asa2021
}
\description{
Numbers-at-age for Sitka Sound herring fishery from ASA 2022-forecast
}
\keyword{datasets}
