% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naa_asa2023_forecast.R
\docType{data}
\name{naa_asa2023_forecast}
\alias{naa_asa2023_forecast}
\title{Numbers-at-age 1976-2022}
\format{
An object of class \code{data.frame} with 47 rows and 7 columns.
}
\usage{
naa_asa2023_forecast
}
\description{
Numbers-at-age for Sitka Sound herring fishery from ASA 2023-forecast
}
\keyword{datasets}
