% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_aub_sim.R
\name{plot_aub_sim}
\alias{plot_aub_sim}
\title{Plot AUB Simulation}
\usage{
plot_aub_sim(B)
}
\arguments{
\item{B}{a data frame returned by the function 'B_sim()'}
}
\description{
This is a wrapper function for plotting the results of average unfished
biomass simulations. AUB (and 25 percent harvest threshold) is computed as the mean
biomass of all years after the cumulative mean series converges.
}
