% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_strata.R
\name{plot_strata}
\alias{plot_strata}
\title{Plot stratification of spawner-recruit data}
\usage{
plot_strata(recruits, spawners, strata = NULL, splits = NULL, years = NULL)
}
\arguments{
\item{recruits}{numeric vector of estimates of annual recruits (millions)}

\item{spawners}{numeric vector of estimates for spawning biomass three years
prior (tonnes)}

\item{strata}{factor vector giving the strata for the empirical spawner-recruit
relationship}

\item{splits}{numeric vector that gives the boundaries of strata. If there are k
unique elements in 'strata' then there should be k-1 elements in 'splits'}

\item{years}{numeric or character vector of years of ASA hindcast estimates,
for plot labeling}
}
\description{
This is a wrapper function for plotting the stratification of spawner-recruit
estimates for use in an empirical spawner-recruit relationship.
}
\references{
\insertRef{carlile1998}{herringSim}
}
