% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_gen.R
\name{R_gen}
\alias{R_gen}
\title{Generate Recruits}
\usage{
R_gen(recruits, spawners, strata = NA, Rbar, sigmahat2, Nt3)
}
\arguments{
\item{recruits}{numeric vector of estimates of annual recruits (millions)}

\item{spawners}{numeric vector of estimates for spawning biomass three years
prior (metric tons)}

\item{strata}{factor vector giving the strata for the empirical spawner-recruit
relationship}

\item{Rbar}{list of mean recruitment for each strata}

\item{sigmahat2}{list of observed standard deviations for each strata}

\item{Nt3}{numeric, spawning biomass in metric tons}
}
\value{
A numeric scalar giving the number of age-3 herring recruits, in millions
}
\description{
This utility function generates age-3 herring recruits from a lognormal
distribution using the observed standard deviations of stratified spawner-
recruit parameter estimates of a herring stock.
}
\details{
Recruitment is treated as a quasi-spawner-independent process. The spawner-
recruit estimates may or may not be stratified. Within each strata, the mean
recruitment multiplied by a log-normal error, where the variance term is
computed from the data within each strata.
}
\references{
\insertRef{methot2011}{herringSim}
}
