% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtruncatednorm.R
\name{rtruncatednorm}
\alias{rtruncatednorm}
\title{Generate Sample from Truncated Normal Distribution}
\usage{
rtruncatednorm(
  n,
  ...,
  lwr_quant = 0.05,
  upr_quant = 0.95,
  lwr = NULL,
  upr = NULL
)
}
\arguments{
\item{n}{integer denoting number of observations in sample}

\item{...}{Arguments passed to \code{rnorm}. Includes \code{mean} and \code{sd}.}

\item{lwr_quant}{Lower limit of normal density}

\item{upr_quant}{Upper limit of normal density}

\item{lwr}{Numeric giving lower limit to truncate past. Supersedes \code{lwr_quant}
if included}

\item{upr}{Numeric giving upper limit to truncate past. Supersedes \code{upr_quant}
if included}
}
\value{
Numeric vector; truncated normal sample of size \code{n}
}
\description{
This function generates a random sample from a pseudo-truncated normal
distribution. Values from a normal distribution that are outside of a user-
specified range are resampled. This occurs until no more values are outside
of the range. The truncated interval may be input as a numeric or as a
normal quantile. The default is to truncate outside of the lower 90\%
of normal density.
}
