% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S_asa1996.R
\docType{data}
\name{S_asa1996}
\alias{S_asa1996}
\title{Survival 1973-1996}
\format{
An object of class \code{data.frame} with 23 rows and 6 columns.
}
\usage{
S_asa1996
}
\description{
Proportional survival estimates from 1973 to 1996 from ASA 1996. Note that
survivals are reported by age class, but are constant across age classes.
}
\keyword{datasets}
