% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S_asa2021.R
\docType{data}
\name{S_asa2021}
\alias{S_asa2021}
\title{Survival 1979-2021}
\format{
An object of class \code{data.frame} with 46 rows and 6 columns.
}
\usage{
S_asa2021
}
\description{
Proportional survival estimates from 1979 to 2021 from ASA 2022-forecast. Note that
survivals are reported by age class, but are constant across age classes.
}
\keyword{datasets}
