% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S_asa2023_forecast.R
\docType{data}
\name{S_asa2023_forecast}
\alias{S_asa2023_forecast}
\title{Survival 1976-2022}
\format{
An object of class \code{data.frame} with 47 rows and 6 columns.
}
\usage{
S_asa2023_forecast
}
\description{
Proportional survival estimates from 1976 to 2022 from ASA 2023-forecast. Note that
survivals are reported by age class, but are constant across age classes.
}
\keyword{datasets}
