% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_recruits.R
\name{sample_recruits}
\alias{sample_recruits}
\title{Sample Recruits}
\usage{
sample_recruits(SSB, sr_list)
}
\arguments{
\item{SSB}{Spawning biomass from which to simulate recruits}

\item{sr_list}{List of stratafied spawner-recruit data, as generated by
\code{\link[=split_sr]{split_sr()}}}
}
\value{
A numeric scalar of millions of recruits
}
\description{
This function simulates future herring age-3 recruits by sampling past recruits.
Future recruits may be generated from a stratified sample using any number
of strata.
}
