% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_sr.R
\name{split_sr}
\alias{split_sr}
\title{Split spawner-recruit into strata}
\usage{
split_sr(recruits, spawners, strata)
}
\arguments{
\item{recruits}{numeric vector of recruits}

\item{spawners}{numeric vector of spawners}

\item{strata}{factor vector delineating strata;}
}
\value{
list of spawners, recruits, and their respective strata.
}
\description{
This utility function splits spawner-recruit estimates into different strata,
listed in ascending order of mean recruit cohort size. All inputs should be
index by the same years
}
