% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sr_asa1996.R
\docType{data}
\name{sr_asa1996}
\alias{sr_asa1996}
\title{ASA 1996 Spawner-Recruit Hindcast}
\format{
An object of class \code{data.frame} with 23 rows and 4 columns.
}
\usage{
sr_asa1996
}
\description{
Parameter estimates of 1996 Sitka herring Age-Structured Assessment (ASA)
hindcast, including estimated recruits (millions), spawning biomass (metric
tons), and strata denoted by Carlile from 1973 to 1996.
}
\keyword{datasets}
