% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sr_asa2021.R
\docType{data}
\name{sr_asa2021}
\alias{sr_asa2021}
\title{ASA 2021 Spawner-Recruit Hindcast}
\format{
An object of class \code{data.frame} with 43 rows and 3 columns.
}
\usage{
sr_asa2021
}
\description{
Parameter estimates from 2022-forecast Sitka herring Age-Structured Assessment (ASA)
hindcast, including estimated recruits (millions) and spawning biomass (metric
tons) from 1979 to 2021.
}
\keyword{datasets}
