% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sr_asa2023_forecast.R
\docType{data}
\name{sr_asa2023_forecast}
\alias{sr_asa2023_forecast}
\title{ASA 2023-forecast Spawner-Recruit parameter estimates}
\format{
An object of class \code{data.frame} with 44 rows and 3 columns.
}
\usage{
sr_asa2023_forecast
}
\description{
Parameter estimates from 2023-forecast Sitka herring Age-Structured Assessment (ASA),
including estimated recruits (millions) and spawning biomass (metric
tons) from 1979 to 2022.
}
\keyword{datasets}
