% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strata_boundaries.R
\name{strata_boundaries}
\alias{strata_boundaries}
\title{Compute strata boundaries}
\usage{
strata_boundaries(sr_list)
}
\arguments{
\item{sr_list}{list of spawner-recruit estimates (and strata) as returned by
the function split_sr}
}
\value{
Vector of spawner biomasses that delineate strata boundaries
}
\description{
This utility function computes the strata boundaries delineating the strata
used in the empirical spawner-recruit relationship. The boundaries are computed
as the mean between the minimum and maximum biomass of adjacent strata.
}
\references{
\insertRef{carlile1998}{herringSim}
}
